# server/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Warubi Platform - FC Köln ITP API
  description: |
    RESTful API for the 1.FC Köln International Talent Program management system.
    
    ## Authentication
    All authenticated endpoints require a Bearer token in the Authorization header.
    Tokens are obtained via the `/api/fckoln/auth/login` endpoint.
    
    ## Multi-Tenancy
    This API supports multiple applications via the `X-App-Key` header.
    Contact support to obtain your application credentials.
    
    ## Rate Limiting
    Requests are rate limited per application. Check response headers for:
    - `X-RateLimit-Limit`: Maximum requests per window
    - `X-RateLimit-Remaining`: Remaining requests in current window
    - `X-RateLimit-Reset`: Unix timestamp when window resets
  version: 1.0.0
  contact:
    name: Warubi Platform Support
    email: support@warubi.io
servers:
  - url: /api/fckoln
    description: FC Köln ITP API
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Invalid request
        correlationId:
          type: string
          format: uuid
          description: Request correlation ID for debugging
...
  responses:
...
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
paths:
...
  /grocery/items:
    get:
      tags: [Grocery]
      summary: List grocery items
      security:
        - BearerAuth: []
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Available grocery items
    post:
      tags: [Grocery]
      summary: Add grocery item
      description: Staff/Admin only
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name, category, price]
              properties:
                name:
                  type: string
                category:
                  type: string
                price:
                  type: number
      responses:
        '201':
          description: Item created
        '403':
          $ref: '#/components/responses/Forbidden'
  /grocery/orders:
    get:
      tags: [Grocery]
      summary: List orders
      description: Players see own orders, staff see all
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Orders list
    post:
      tags: [Grocery]
      summary: Place order
      description: Subject to weekly budget limit (30 EUR)
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [deliveryDate, items]
              properties:
                deliveryDate:
                  type: string
                  format: date
                items:
                  type: array
                  items:
                    type: object
                    required: [itemId, quantity]
                    properties:
                      itemId:
                        type: integer
                      quantity:
                        type: integer
                        minimum: 1
      responses:
        '201':
          description: Order placed
        '400':
          description: Budget exceeded or invalid items
  /grocery/orders/{id}:
    get:
      tags: [Grocery]
      summary: Get order details
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Order with items
        '403':
          description: Not authorized to view this order
        '404':
          $ref: '#/components/responses/NotFound'